% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiCall.R
\name{apiCall}
\alias{apiCall}
\title{apiCall - Base call to Core REST API.}
\usage{
apiCall(coreApi, body, encode, special = NULL, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{body}{body for request}

\item{encode}{encodeing to use for request option are "multipart", "form", "json", "raw"}

\item{special}{- pased to buildUrl for special sdk endpoints}

\item{useVerbose}{Use verbose communitcation for debugging}
}
\value{
RETURN return the entire http response
}
\description{
\code{apiCall}  base call to Core REST API.
}
\details{
\code{apiCall}  base call to Core REST API.
}
\examples{
\dontrun{
api<-CoreAPI("PATH TO JSON FILE")
login<- CoreAPI::authBasic(api)
response <-CoreAPI::apiCall(login$coreApi,body,"json",,special=NULL,useVerbose=FALSE)
logOut(login$coreApi )
}
}
\author{
Craig Parman
}

