% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachFile.R
\name{attachFile}
\alias{attachFile}
\title{attachFile - attaches a file to an entitiy or attribute.}
\usage{
attachFile(coreApi, barcode, filename, filepath, targetAttributeName = "",
  useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{barcode}{User provided barcode as a character string}

\item{filename}{name to use for the attached file}

\item{filepath}{path to the file to attach}

\item{targetAttributeName}{- if included the name if the attribute to attach the file to.  Must be in all caps.}

\item{useVerbose}{Use verbose communitcaion for debuggins}
}
\value{
RETURN returns a list $entity contains entity information, $response contains the entire http response
}
\description{
\code{attachFile} Attaches a file to entity identified by barcode or one ot its attributes.
}
\details{
\code{attachFile}  attaches a file to an entitiy or attribute.
}
\examples{
\dontrun{
api<-CoreAPI("PATH TO JSON FILE")
login<- CoreAPI::authBasic(api)
newitem<-CoreAPI::attachFile(response$coreApi,barcode,filename,filepath,targetAttributeName="",useVerbose=FALSE)
logOut(login$coreApi )
}
}
\author{
Craig Parman
}

