% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authBasic.R
\name{authBasic}
\alias{authBasic}
\title{Authenticates against the LIMS using basic authentication}
\usage{
authBasic(coreUrl, user, pwd, account = NULL, useVerbose = FALSE)
}
\arguments{
\item{coreUrl}{character string that is the url of LIMS}

\item{user}{character string that is username}

\item{pwd}{character string that is password}

\item{account}{character string that is the account to log into for multi tenant systems}

\item{useVerbose}{TRUE or FALSE to indicate if verbose options should be used in http POST}
}
\value{
returns a list $sessionInfo contains jsessionid and employeeid, $response contains the entire http response
}
\description{
\code{logOut} Logs in and create a jsession using a username and password..
}
\details{
\code{authBasic} authenticates to Core API
}
\examples{
\dontrun{
response<- CoreAPI::authBasic(coreUrl,user,pwd)
js<-response$sessionInfo$jessionid
logOut(coreUrl,js,postVerbose=FALSE )}
}
\author{
Craig Parman
}

