% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildODATAUrl.R
\name{buildODATAUrl}
\alias{buildODATAUrl}
\title{buildODATAUrl - build URL for call to Core ODATA API.}
\usage{
buildODATAUrl(coreApi, resource = NULL, query = NULL, special = NULL,
  useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{resource}{resource path (required except for special requests)}

\item{query}{and additional property options (optional)}

\item{special}{flag for special sdk endpoints}

\item{useVerbose}{Use verbose communication for debugging}
}
\value{
RETURN Core REST URL
}
\description{
\code{buildODATAUrl} build URL for call to Core REST API.
}
\details{
\code{buildODATAUrl}  build URL for call to Core ODATA API.
}
\examples{
\dontrun{
api<-CoreAPI("PATH TO JSON FILE")
login<- CoreAPI::authBasic(api)
response <-CoreAPI::buildODATAUrl(coreApi,"Sample","('PS1')")
logOut(login$coreApi )
}
}
\author{
Craig Parman ngsAnalytics, ngsanalytics.com
}
