% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildUrl.R
\name{buildUrl}
\alias{buildUrl}
\title{buildUrl - build URL for call to Core REST API.}
\usage{
buildUrl(coreApi, special = NULL, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{special}{flag for special sdk endpoints}

\item{useVerbose}{Use verbose communication for debugging}
}
\value{
RETURN Core REST URL
}
\description{
\code{buildUrl} build URL for call to Core REST API.
}
\details{
\code{apiCall}  base call to Core REST API.
}
\examples{
\dontrun{
api<-CoreAPI("PATH TO JSON FILE")
login<- CoreAPI::authBasic(api)
response <-CoreAPI::apiCall(login$coreApi,body,"json",useVerbose=FALSE)
logOut(login$coreApi )
}
}
\author{
Craig Parman
}

