% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreAPI.R
\name{coreAPI}
\alias{coreAPI}
\title{Creates a object of class coreAPI that contains connection information
\code{etradeAPI("path to json")}}
\usage{
coreAPI(CoreAccountInfo)
}
\arguments{
\item{CoreAccountInfo}{file with account information in json format.}
}
\value{
Object of class coreAPI
}
\description{
Creates a object of class coreAPI that contains connection information
\code{etradeAPI("path to json")}
}
\details{
{ Creates a object of class coreAPI that contains user name, \cr
         password base url, and account \cr
         if needed. It has slots for account, jsessionID, AWSELB, and base URL. \cr
         Requires a json file that contains the user pwd and account. \cr
          }
\code{#'Creates a object of class coreAPI that contains account information}
\code{coreAPI("path to json")}
\code{
           [{
             "user": "xxxxxxxxx",         \cr
             "pwd": "xxxxxxxxx",          \cr
             "coreUrl": "xxxxxxx",        \cr
             "account": "xxxxxxxxx",      \cr
           }]                             \cr
}
}
\examples{
\dontrun {  }
}

