% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEntity.R
\name{createEntity}
\alias{createEntity}
\title{createEntity - Create a new instance of a entitiy.}
\usage{
createEntity(coreApi, entityType, attributeValues = NULL, locationId = NULL,
  projectIds = NULL, barcode = NULL, associations = NULL,
  useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{entityType}{entity type to get as character string}

\item{attributeValues}{atributes as a list of key-vlaues pairs}

\item{locationId}{location ID for inital location as character string}

\item{projectIds}{project comma seperated list of project IDs as character string}

\item{barcode}{User provided barcode as a character string}

\item{associations}{association as a list of dataframes (see details)}

\item{useVerbose}{Use verbose communitcaion for debuggins}
}
\value{
RETURN returns a list $entity contains entity information, $response contains the entire http response
}
\description{
\code{createEntity} Creates a new entity instance. Required inputs are url, jsessionId and entityType.
}
\details{
\code{createEntity} Creates a new entity instance.
}
\examples{
\dontrun{
api<-CoreAPI("PATH TO JSON FILE")
login<- CoreAPI::authBasic(api)
item<-createEntityt(login$coreApi,"Entity_Type")
logOut(login$coreApi )
}
}
\author{
Craig Parman
}

