% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createExperimentSample.R
\name{createExperimentSample}
\alias{createExperimentSample}
\title{createExperimentSample- Creates an experiment sample.}
\usage{
createExperimentSample(coreApi, entityType, sampleLotBarcode, experimentBarcode,
  attributeValues = NULL, locationId = NULL, projectIds = NULL,
  barcode = NULL, associations = NULL, concentration = NULL,
  concentrationUnit = NULL, timeMin = NULL, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{entityType}{entity type to get as character string}

\item{sampleLotBarcode}{parent sample barcode}

\item{experimentBarcode}{experiment barcode}

\item{attributeValues}{atributes as a list of key-vlaues pairs}

\item{locationId}{location ID for inital location as character string}

\item{projectIds}{project comma seperated list of project IDs as character string}

\item{barcode}{User provided barcode as a character string}

\item{associations}{association as a list of dataframes (see details)}

\item{concentration}{sample lot concentration}

\item{concentrationUnit}{concentration unit}

\item{timeMin}{min time value}

\item{useVerbose}{Use verbose communitcaion for debuggins}
}
\value{
RETURN returns a list $entity contains entity information, $response contains the entire http response
}
\description{
\code{createExperimentSample} Creates a new sample lot using the parent sample barcode
}
\details{
\code{createExperimentSample} Creates an experiment sample.
}
\examples{
\dontrun{
api<-CoreAPI("PATH TO JSON FILE")
login<- CoreAPI::authBasic(api)
item<-createExperimentSample(login$coreApi,
logOut(login$coreApi )
}
}
\author{
Craig Parman
}

