% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSampleLot.R
\name{createSampleLot}
\alias{createSampleLot}
\title{createEntity - Create a sample lot.}
\usage{
createSampleLot(coreApi, entityType, sampleBarcode, attributeValues = NULL,
  locationId = NULL, projectIds = NULL, barcode = NULL,
  associations = NULL, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{entityType}{entity type to get as character string}

\item{sampleBarcode}{parent sample barcode}

\item{attributeValues}{atributes as a list of key-vlaues pairs}

\item{locationId}{location ID for inital location as character string}

\item{projectIds}{project comma seperated list of project IDs as character string}

\item{barcode}{User provided barcode as a character string}

\item{associations}{association as a list of dataframes (see details)}

\item{useVerbose}{Use verbose communitcaion for debuggins}
}
\value{
RETURN returns a list $entity contains entity information, $response contains the entire http response
}
\description{
\code{createSampleLot} Creates a new sample lot using the parent sample barcode
}
\details{
\code{createSampleLot} Creates a lot of a sample.
}
\examples{
\dontrun{
api<-CoreAPI("PATH TO JSON FILE")
login<- CoreAPI::authBasic(api)
item<-createSampleLot(login$coreApi,"Sample_Lot_Name","sample_barcode"
logOut(login$coreApi )
}
}
\author{
Craig Parman
}

