% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentUnPublish.R
\name{experimentUnPublish}
\alias{experimentUnPublish}
\title{experimentUnPublish Unpublishes an experiment.}
\usage{
experimentUnPublish(coreApi, experimentType, experimentBarcode,
  useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{experimentType}{experiment entity type}

\item{experimentBarcode}{barcode of the experiment}

\item{useVerbose}{Use verbose communication for debugging}
}
\value{
RETURN returns a list $entity contains updated experiment information, $response contains the entire http response
}
\description{
\code{experimentUnPublish} -  Unpublishes an experiment.
}
\details{
\code{experimentUnPublish} Unpublishes an experiment.
}
\examples{
\dontrun{
api<-CoreAPI("PATH TO JSON FILE")
login<- CoreAPI::authBasic(api)
update<- CoreAPI::experimentUnPublish(login$coreApi,experimentType, exptbarcode,useVerbose = TRUE)
logOut(login$coreApi )
}
}
\author{
Craig Parman
}
