% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContainerLineage.R
\name{getContainerLineage}
\alias{getContainerLineage}
\title{Get lineage for a container by barcode}
\usage{
getContainerLineage(coreApi, barcode, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{barcode}{barcode of container to get lineage for}

\item{useVerbose}{TRUE or FALSE to indicate if verbose options should be used in http POST}
}
\value{
returns a list $entity contains entity information, $response contains the entire http response
}
\description{
\code{getContainerLineage} Get the parent and child conttainers for a container referenced by barcode..
}
\details{
\code{getContainerLineage} get an entity from the LIMS by barcode and
}
\examples{
\dontrun{
api<-CoreAPI("PATH TO JSON FILE")
login<- CoreAPI::authBasic(api)
lineage<-getContainerLineage(login$coreApi,"barcode")
parents <- ineage$entity$parents
logOut(login$coreApi )
}
}
\author{
Craig Parman
}

