% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEntityById.R
\name{getEntityById}
\alias{getEntityById}
\title{getEntityById - Get an entity from the LIMS by entity ID.}
\usage{
getEntityById(coreApi, entityType, entityId, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{entityType}{entity type to get}

\item{entityId}{ID of entity to get}

\item{useVerbose}{TRUE or FALSE to indicate if verbose options should be used in http POST}
}
\value{
returns a list $entity contains entity information, $response contains the entire http response
}
\description{
\code{getEntityById} Get an entity from the LIMS by ID and entyType.
}
\details{
\code{getEntityById} get an entity from the LIMS by ID.
}
\examples{
\dontrun{
api<-CoreAPI("PATH TO JSON FILE")
login<- CoreAPI::authBasic(api)
item<-getEntityById(login$coreApi,"entityType","entityId")
logOut(login$coreApi)
}
}
\author{
Craig Parman
}

