% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEntityByName.R
\name{getEntityByName}
\alias{getEntityByName}
\title{getEntityByName - Get an entity from the LIMS by name.}
\usage{
getEntityByName(coreApi, entityType, name, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{entityType}{entity type to get}

\item{name}{name of entity to get}

\item{useVerbose}{TRUE or FALSE to indicate if verbose options should be used in http POST}
}
\value{
returns a list $entity contains entity information, $response contains the entire http response
}
\description{
\code{getEntityByName} Get an entity from the LIMS by name and entityType.
}
\details{
\code{getEntityByName} get an entity from the LIMS by name
}
\examples{
\dontrun{
api<-CoreAPI("PATH TO JSON FILE")
login<- CoreAPI::authBasic(api)
item<-getEntityByName(login$coreApi,"entityType","name")
logOut(login$coreApi)
}
}
\author{
Craig Parman
}

