% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExperimentSamples.R
\name{getExperimentSamples}
\alias{getExperimentSamples}
\title{getExperimentSamples - Gets experiment and experiment samples from experiment identified by barcode.}
\usage{
getExperimentSamples(coreApi, entityType, barcode, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{entityType}{experiment entity type to get}

\item{barcode}{barcode of entity to get}

\item{useVerbose}{TRUE or FALSE to indicate if verbose options should be used in http POST}
}
\value{
returns a list $entity contains entity information, $response contains the entire http response
}
\description{
\code{ getExperimentSamples}  Gets experiment and experiment samples from experiment identified by barcode.
}
\details{
\code{getExperimentSamples}  Gets experiment and experiment samples from experiment identified by barcode.
}
\examples{
\dontrun{
api<-CoreAPI("PATH TO JSON FILE")
login<- CoreAPI::authBasic(api)
experiment<- getExperimentSamples(login$coreApi,"entityType","barcode")
logOut(login$coreApi)
}
}
\author{
Craig Parman
}

