% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logOut.R
\name{logOut}
\alias{logOut}
\title{Log user out of the LIMS}
\usage{
logOut(coreUrl, jsessionid, useVerbose = FALSE)
}
\arguments{
\item{coreUrl}{base url of LIMS}

\item{jsessionid}{jsessionid to log out of username}

\item{useVerbose}{use verbose option for debuggin in http POST}
}
\value{
returns list with $success = TRUE when sucessful, $response contains the entire http response
}
\description{
\code{logOut} logs out of the current jsession.
}
\details{
\code{logOut} logs user out of the LIMS using Core API
}
\examples{
\dontrun{
response<- CoreAPI::authBasic(coreUrl,user,pwd)
js<-response$sessionInfo$jessionid
logOut(coreUrl,js,postVerbose=FALSE )}
}
\author{
Craig Parman
}

