% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODATAcleanName.R
\name{ODATAcleanName}
\alias{ODATAcleanName}
\title{ODATAcleanName - converts names to ODATA compliant version. Used to clean names in ODATA calls.}
\usage{
ODATAcleanName(name)
}
\arguments{
\item{name}{string to clean}
}
\value{
Returns name in ODATA compliant form
}
\description{
\code{ODATAcleanName} - converts names to ODATA compliant version. Used to clean names in ODATA calls.
}
\details{
\code{ODATAcleanName} Clean a name for ODATA.
}
\examples{
\dontrun{
new_name <-CoreAPIV2::ODATAcleanName("384 Well Plate")
new_name
_384_WELL_PLATE
 }
}
\author{
Craig Parman ngsAnalytics, ngsanalytics.com
}
