% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transferCellContents.R
\name{transferCellContents}
\alias{transferCellContents}
\title{transferCellContents -  Transfers contents from one cell to another.}
\usage{
transferCellContents(coreApi, sourceCellID, destCellID, amount, concentration,
  amountUnit, concentrationUnit, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{sourceCellID}{source cell ID}

\item{destCellID}{destination cell ID}

\item{amount}{amount to transfer}

\item{concentration}{concentration for destination cell}

\item{amountUnit}{valid units for amount}

\item{concentrationUnit}{valid units for concentration}

\item{useVerbose}{use verbode messaging for debugging}
}
\value{
RETURN returns a list $entity contains cell information, $response contains the entire http response
}
\description{
\code{transferCellContents} - Get information about container cell contents.
}
\details{
\code{transferCellContents} puts a cell lot in a container cell
}
\examples{
\dontrun{
api<-CoreAPI("PATH TO JSON FILE")
login<- CoreAPI::authBasic(api)
newcell<-transferCellContents(login$coreApi,"98811","93221","1",".1","mg/ML","uL")
logOut(login$coreApi )
}
}
\author{
Craig Parman
}

