% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCellContents.R
\name{updateCellContents}
\alias{updateCellContents}
\title{updateCellContents -  Puts a cell lot in a container cell.}
\usage{
updateCellContents(coreApi, containerType, containerBarcode, containerCellNum,
  sampleLotBarcode, amount, amountUnit, concentration, concentrationUnit,
  useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{containerType}{container entity type}

\item{containerBarcode}{container barcode}

\item{containerCellNum}{container cell number}

\item{sampleLotBarcode}{barcode of lot to add to cell}

\item{amount}{amount to add (numeric)}

\item{amountUnit}{units}

\item{concentration}{(numeric)}

\item{concentrationUnit}{concentration units}

\item{useVerbose}{use verbose communications for debugging}
}
\value{
RETURN returns a list $entity contains updated container
        information, $response contains the entire http response
}
\description{
\code{updateCellContents} - Puts a cell lot in a container cell.
}
\details{
\code{updateCellContents} Puts a cell lot in a container cell
}
\examples{
\dontrun{
api<-CoreAPI("PATH TO JSON FILE")
login<- CoreAPI::authBasic(api)
cell<-
logOut(login$coreApi )
}
}
\author{
Craig Parman
}

