% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateExperimentContainers.R
\name{updateExperimentContainers}
\alias{updateExperimentContainers}
\title{updateExperimentContainers -  Adds a container to an experiment.}
\usage{
updateExperimentContainers(coreApi, containerBarcode, experimentType,
  experimentBarcode, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{containerBarcode}{barcode of container that has a cell lot in it}

\item{experimentType}{experiment entity type}

\item{experimentBarcode}{barcode of the experiment}

\item{useVerbose}{Use verbose communication for debugging}
}
\value{
RETURN returns a list $entity contains updated experiment information,
        $response contains the entire http response
}
\description{
\code{updateExperimentContainers} - Adds a container to an experiment.
}
\details{
\code{updateExperimentContainers} Adds a container to an experiment.
}
\examples{
\dontrun{
api<-CoreAPI("PATH TO JSON FILE")
login<- CoreAPI::authBasic(api)
update<- CoreAPI::updateExperimentContainers(login$coreApi,containerBarcode,
                                            experimentType, newExptbarcode,useVerbose = TRUE)
                                            logOut(login$coreApi )
}
}
\author{
Craig Parman
}

