% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateExperimentSampleData.R
\name{updateExperimentSampleData}
\alias{updateExperimentSampleData}
\title{updateExperimentSampleData - Upate experiment sample data.}
\usage{
updateExperimentSampleData(coreApi, entityType, experimentSamplebarcode,
  assayAttributeValues, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{entityType}{entity type to get as character string}

\item{experimentSamplebarcode}{User provided barcode as a character string}

\item{assayAttributeValues}{assay attributes as a list of key-values pairs}

\item{useVerbose}{Use verbose communication for debugging}
}
\value{
RETURN returns a list $entity contains entity information, $response contains the entire http response
}
\description{
\code{updateExperimentSampleData} Update experiment sample data.
}
\details{
\code{updateExperimentSampleData} Upate experiment sample data.
}
\examples{
\dontrun{
api<-CoreAPI("PATH TO JSON FILE")
login<- CoreAPI::authBasic(api)
newdata<-CoreAPI::updateExperimentSampleData(login$coreApi,entityType,experimentSampleBarcode,assayAtributeValues)
logOut(login$coreApi ) response<- CoreAPI::authBasic(coreApi)
}
}
\author{
Craig Parman
}

