% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attachFile.R
\name{attachFile}
\alias{attachFile}
\title{attachFile - attaches a file to an entitiy or attribute.}
\usage{
attachFile(coreUrl, jessionid, barcode, filename, filepath,
  targetAttributeName = "", useVerbose = FALSE)
}
\arguments{
\item{coreUrl}{character string that is the url of LIMS}

\item{jessionid}{valid jesssionid as character string}

\item{barcode}{User provided barcode as a character string}

\item{filename}{name to use for the attached file}

\item{filepath}{path to the firl to attach}

\item{targetAttributeName}{- if included the name if the attribute to attach the file to.  Must be in all caps.}

\item{useVerbose}{Use verbose communitcaion for debuggins}
}
\value{
RETURN returns a list $entity contains entity information, $response contains the entire http response
}
\description{
\code{attachFile} Attaches a file to entity identified by barcode or one ot its attributes.
}
\details{
\code{attachFile} Creates a new entity instance.
}
\examples{
\dontrun{
response<- CoreAPI::authBasic(coreUrl,user,pwd)
jessionid<-response$sessionInfo$jessionid
newitem<-CoreAPI::fileAttach(coreUrl,jessionid,entityType,barcodefilename,filepath,targetAttributeName="",useVerbose=FALSE)
jsonlite::toJSON(newitem$item)
logOut(coreUrl,js,postVerbose=FALSE )
}
}
\author{
Craig Parman
}

