% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authBasic.R
\name{authBasic}
\alias{authBasic}
\title{Authenticates against the LIMS using basic authentication}
\usage{
authBasic(coreApi, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{object of class coreApi that contains user, password,  baseURL and \cr
account. account is required if user has access to multiple tenants.}

\item{useVerbose}{- Use verbose settings for HTTP commands}
}
\value{
returns a list with coreApi which returns the passed coreApi object with  jsessionid, \cr
           awselb and employeeid populated, $response contains the entire http response \cr
}
\description{
\code{authBasic} Logs in and returns a fully populated coreApi object.
}
\details{
\code{authBasic} authenticates to Core API
}
\examples{
\dontrun{
response<- CoreAPI::authBasic(coreAPI)
}
}
\author{
Craig Parman
}

