% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authBasic.R
\name{authBasic}
\alias{authBasic}
\title{Authenticates against the LIMS using basic authentication}
\usage{
authBasic(coreApi, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{object of class coreApi that contains user, password,  baseURL and account(optional)}

\item{useVerbose}{- Use verbose setting for HTTP commands}
}
\value{
{ returns a list $coreApi which retuned passed coreApi object with  jsessionid, \cr
           awselb and employeeid populated, $response contains the entire http response \cr
        }
}
\description{
\code{logOut} Logs in and create a jsession using a username and password..
}
\details{
\code{authBasic} authenticates to Core API
}
\examples{
\dontrun{
response<- CoreAPI::authBasic(coreUrl,user,pwd)
js<-response$sessionInfo$jessionid
logOut(coreUrl,js,postVerbose=FALSE )}
}
\author{
Craig Parman
}

