% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreAPI.R
\name{coreAPI}
\alias{coreAPI}
\title{coreAPI Creates a object of class coreAPI that contains user and connection information.}
\usage{
coreAPI(CoreAccountInfo)
}
\arguments{
\item{CoreAccountInfo}{file with account information in json format.}
}
\value{
Object of class coreAPI
}
\description{
coreAPI Creates a object of class coreAPI that contains user and connection information.
}
\details{
{ Creates a object of class coreAPI that contains user name,
         password base url, account, port.
         It has slots for account, jsessionId, AWSELB, and base URL.
         Requires a json file that is a POSTMAN environment file.
        \code{#'Creates a object of class coreAPI that contains account information}
        \code{coreAPI("path to json")}.
}
The json must include the fields shown below.
    \code{
     {
   "values": [
 {
   "key": "tenant",
   "value": "R-Integration_Baseline"
 },
 {
   "key": "TenantShortName",
   "value": "bp2",
   "type": "text",
   "enabled": true
 },
 {
   "key": "scheme",
   "value": "https"
 },
 {
   "key": "host",
   "value": "lims.ccc.cloud"
 },
 {
   "key": "username",
  "value": "yyyy"
 },
 {
   "key": "password",
   "value": "xxxxx"
 },
 {
   "key": "port",
   "value": "443"
 }
 ]

}


    }

 The account value may be set to "" if the user only has access to one tenant.
As an alternative the environment json object from Postman can be used directly.
}
\examples{
\dontrun{
api<-CoreAPIV::coreAPI("/home/environment.json")
}
}
\author{
Craig Parman ngsAnalytics, ngsanalytics.com
}
