% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreAPI.R
\name{coreAPI}
\alias{coreAPI}
\title{Creates a object of class coreAPI that contains user and connection information
\code{coreApi("path to json")}}
\usage{
coreAPI(CoreAccountInfo)
}
\arguments{
\item{CoreAccountInfo}{file with account information in json format.}
}
\value{
Object of class coreAPI
}
\description{
Creates a object of class coreAPI that contains user and connection information
\code{coreApi("path to json")}
}
\details{
Creates a object of class coreAPI that contains user name,            \cr
         password base url, and account                                        \cr
         if needed. It has slots for account, jsessionID, AWSELB, and base URL. \cr
         Requires a json file that contains the user pwd and account.           \cr

        \code{#'Creates a object of class coreAPI that contains account information} \cr
        \code{coreAPI("path to json")}  \cr \cr
The json must include the fields shown below.  The account value may be set to "" if the user only has access to one tenant. \cr \cr
Example json object.  \cr \cr
         \code{
               [{                             \cr
                 "user": "xxxxxxxxx",         \cr
                 "pwd": "xxxxxxxxx",          \cr
                 "coreUrl": "xxxxxxx",        \cr
                "account": "xxxxxxxxx",       \cr
               }]                              \cr
          }
}
\examples{
\dontrun{
  api<-coreApi("/home")
}
}

