% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreAPI.R
\name{coreAPI}
\alias{coreAPI}
\title{coreAPI - Creates a object of class coreAPI that contains user and connection information.}
\usage{
coreAPI(CoreAccountInfo)
}
\arguments{
\item{CoreAccountInfo}{file with account information in json format.}
}
\value{
Object of class coreAPI
}
\description{
coreAPI - Creates a object of class coreAPI that contains user and connection information.
}
\details{
{ Creates a object of class coreAPI that contains user name,
         password base url, and account
         if needed. It has slots for account, jsessionID, AWSELB, and base URL.
         Requires a json file that contains the user pwd and account.
        \code{#'Creates a object of class coreAPI that contains account information}
        \code{coreAPI("path to json")}.

The json must include the fields shown below.  The account value may be set to ""
if the user only has access to one tenant.
Example json object.

         \code{

               [{
                 "user": "xxxxxxxxx",
                 "pwd": "xxxxxxxxx",
                 "coreUrl": "xxxxxxx",
                "account": "xxxxxxxxx",
               }]
          }
}
}
\examples{
\dontrun{
api<-coreApi("/home")
}
}

