% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getByBarcodeEntity.R
\name{getByBarcodeEntity}
\alias{getByBarcodeEntity}
\title{Get an entity from the LIMS by barcode}
\usage{
getByBarcodeEntity(coreUrl, jsessionid, entityType, barcode,
  useVerbose = FALSE)
}
\arguments{
\item{coreUrl}{character string that is the url of LIMS}

\item{jsessionid}{valid jsessionid}

\item{entityType}{entity type to get}

\item{barcode}{barcode of entity to get}

\item{useVerbose}{TRUE or FALSE to indicate if verbose options should be used in http POST}
}
\value{
returns a list $entity contains entity information, $response contains the entire http response
}
\description{
\code{getByBarcodeEntity} Get an entity from the LIMS by barcode and entyType.
}
\details{
\code{getByBarcodeEntity} get an entity from the LIMS by barcode and
}
\examples{
\dontrun{
response<- CoreAPI::authBasic(coreUrl,user,pwd)
js<-response$sessionInfo$jsessionid
item<-CoreAPI::getByBarcodeEntity(coreUrl,js,entityType,barcode,useVerbose)
jsonlite::toJSON(item$entity)
logOut(coreUrl,js,postVerbose=FALSE )}
}
\author{
Craig Parman
}

