% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logOut.R
\name{logOut}
\alias{logOut}
\title{Log user out of the LIMS}
\usage{
logOut(coreApi, useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object returned during log in}

\item{useVerbose}{use verbose option for debuggin in http POST}
}
\value{
returns list with $success = TRUE when sucessful, $response contains the entire http response
}
\description{
\code{logOut} logs out of the current jsession.
}
\details{
\code{logOut} logs user out of the LIMS using Core API
}
\examples{
\dontrun{
api<-CoreAPI("PATH TO JSON FILE")
response<- CoreAPI::authBasic(api)
logOut(response$coreApi,useVerbose=TRUE )
}
}
\author{
Craig Parman
}

