% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateEntity.R
\name{updateEntity}
\alias{updateEntity}
\title{updateEntity - Updates an instance of a entity.}
\usage{
updateEntity(coreApi, entityType, barcode, attributeValues = NULL,
  locationId = NULL, projectIds = NULL, associations = NULL,
  useVerbose = FALSE)
}
\arguments{
\item{coreApi}{coreApi object with valid jsessionid}

\item{entityType}{entity type to get as character string}

\item{barcode}{User provided barcode as a character string}

\item{attributeValues}{attributes as a list of key-values pairs}

\item{locationId}{location ID for initial location as character string}

\item{projectIds}{project comma separated list of project IDs as character string}

\item{associations}{association as a list of dataframes (see details)}

\item{useVerbose}{Use verbose communication for debugging}
}
\value{
RETURN returns a list $entity contains entity information, $response contains the entire http response
}
\description{
\code{updateEntity} Updates an instance of a entity.
}
\details{
\code{updateEntity}  Updates an instance of a entity.
}
\examples{
\dontrun{
api<-CoreAPI("PATH TO JSON FILE")
login<- CoreAPI::authBasic(api)
newitem<-CoreAPI::updateEntity(login$coreApi,entityType,barcode)
logOut(login$coreApi ) response<- CoreAPI::authBasic(coreApi)
}
}
\author{
Craig Parman
}

